Ported Backend Selection from Demo Version (#24)
## Technical Description

Currently, icondusk-e2e uses the demo branch to be able to select the backend. However, since we want to test use our new dusk features in the e2e repo, this should change to master. Hence, this PR to port the -b flag introduced in the demo branch also to master, since it's required by the end to end repo.